VSP-558 [iOS] Migrate to iOS 15
iOS 15 extended the use of scrollEdgeAppearance. When running on apps that use iOS 14 or earlier, this property applies to navigation bars with large titles. In iOS 15, this property applies to all navigation bars.

Switched to the new UINavigationBarAppearance for configuration on all systems iOS 13+, since the legacy method is yielding unstable results.