VSP-581 iOS: member role selection scroll wheel should not auto save
Improved the usability of dropdowns:
If there is no value selected, then we automatically select the first value in the list, so that the user doesn't have to move to a new value then back again to the first one.
When picking a value in the selection wheel, we do not automatically assume it's the correct one and the wheel stays visible.